refactor: move fuel_moisture module to dmoist crate
move the `fuel_moisture.rs` module to the `dmoist` crate. Keep. `create_terrain_with_fuel_moisture` in `firelib` though, so `dmoist` does not depend on `firelib`

```git-revs
5b975b1  (Base revision)
953b2e4  Add HourlyMoistureResults struct to dmoist crate
f7de0cc  Snapshot before codemcp change
ced2572  Snapshot before codemcp change
69e13a2  Snapshot before codemcp change
90fc614  Fix variable name in assignment
770bb53  Snapshot before codemcp change
4b3f644  Fix 10hr assignment variable name
d82cf39  Snapshot before codemcp change
b306861  Fix 100hr assignment variable name
65d176d  Snapshot before auto-test
34759d7  Update firelib fuel_moisture.rs to import from dmoist and keep only create_terrain_with_fuel_moisture
HEAD     Add test for calculate_hourly_fuel_moisture function
```

codemcp-id: 10-refactor-move-fuel-moisture-module-to-dmoist-crate
