+++ crates/dmoist/src/lib.rs
@@ -1,8 +1,136 @@
 #![feature(import_trait_associated_functions)]
 #![no_std]
+use num::traits::FloatConst;
 #[allow(unused_imports)]
 use num::Float;
 use num::Float::nan;
+
+/// Output data for hourly fuel moisture calculation
+#[derive(Debug, Clone)]
+pub struct HourlyMoistureResults {
+    /// 1-hour fuel moisture values for each hour [0-23]
+    pub d1hr: [f32; 24],
+    /// 10-hour fuel moisture values for each hour [0-23]
+    pub d10hr: [f32; 24],
+    /// 100-hour fuel moisture values for each hour [0-23]
+    pub d100hr: [f32; 24],
+}
+
+// Calculate hourly fuel moisture values using dmoist algorithms
+//
+// This function takes meteorological inputs and returns calculated fuel moisture
+// values for each hour of the day (0-23) that can be used with TerrainCuda structures.
+//
+// # Arguments
+// * `temperature` - Hourly temperature in degrees Celsius [0-23]
+// * `humidity` - Hourly relative humidity in percentage [0-23]
+// * `cloud_cover` - Hourly cloud cover in percentage [0-23]
+// * `slope` - Fixed slope ratio
+// * `aspect` - Fixed aspect value in radians
+// * `precipitation_6_days` - Daily total precipitation for the last 6 days in mm [day-6 to day-1]
+// * `month` - Month (1-12) for seasonal adjustments
+// * `fuel_model` - Fuel model code (0-13) for model-specific adjustments
+//
+// # Returns
+// * `HourlyMoistureResults` containing d1hr, d10hr, and d100hr values for each hour
+//
+// # Example
+// ```no_run
+// use dmoist::calculate_hourly_fuel_moisture;
+//
+// let temperature = [15.0; 24]; // 15°C for all hours
+// let humidity = [60.0; 24];    // 60% RH for all hours
+// let cloud_cover = [50.0; 24]; // 50% cloud cover for all hours
+// let slope = 0.577;            // 30 degree slope as ratio (tan(30°))
+// let aspect = 3.14159;         // South-facing (π radians)
+// let precipitation_6_days = [0.0, 2.0, 0.0, 5.0, 0.0, 1.0]; // mm for last 6 days
+// let month = 6;                // June
+//
+// let results = calculate_hourly_fuel_moisture(
+//     &temperature, &humidity, &cloud_cover, slope, aspect,
+//     &precipitation_6_days, month, 2
+// );
+// ```
+#[allow(clippy::too_many_arguments)]
+pub fn calculate_hourly_fuel_moisture(
+    temperature: &[f32; 24],
+    humidity: &[f32; 24],
+    cloud_cover: &[f32; 24],
+    slope: f32,
+    aspect: f32,
+    precipitation_6_days: &[f32; 6],
+    month: i32,
+    fuel_model: i32,
+) -> HourlyMoistureResults {
+    let mut d1hr_results = [0.0; 24];
+    let mut d10hr_results = [0.0; 24];
+    let mut d100hr_results = [0.0; 24];
+
+    // Calculate maximum precipitation effect for the 6-day period
+    let max_precipitation_effect = efecto_precipitacion_maximo(month, precipitation_6_days);
+
+    for (hour, d1hr_value) in d1hr_results.iter_mut().enumerate() {
+        // Calculate base humidity index (HCB)
+        let hcb_base = hcb(temperature[hour], humidity[hour], hour as i32) as f32;
+
+        // Correct HCB for terrain and cloud shading effects
+        let hcb_corrected = corrige_hcb_por_sombreado(
+            hcb_base,
+            cloud_cover[hour],
+            month,
+            aspect * 180.0 / f32::PI(),       // Convert radians to degrees
+            slope.atan() * 180.0 / f32::PI(), // Convert ratio to degrees
+            fuel_model,
+            hour as i32,
+        );
+
+        // Calculate ignition probability without precipitation correction
+        let prob_ignition_uncorrected = probabilidad_ignicion(
+            temperature[hour],
+            cloud_cover[hour],
+            hcb_corrected,
+            fuel_model,
+        );
+
+        // Correct ignition probability for precipitation effects
+        let prob_ignition_corrected =
+            corrige_prob_por_pp(prob_ignition_uncorrected, max_precipitation_effect);
+
+        // Calculate 1-hour fuel moisture
+        *d1hr_value = d1hr(
+            hcb_corrected,
+            prob_ignition_uncorrected,
+            prob_ignition_corrected,
+        );
+    }
+
+    // For 10hr and 100hr, we need previous values for the calculation
+    // For now, we'll use a simplified approach and calculate them based on d1hr
+    // In a complete implementation, these would need historical values
+    for (hour, d10hr_value) in d10hr_results.iter_mut().enumerate() {
+        // Calculate 10-hour moisture using the dmoist algorithm
+        // We need values from 6 and 15 hours ago, but for simplicity we'll use current and previous day estimates
+        let d1hr_6h = d1hr_results[6];
+        let d1hr_15h = d1hr_results[15];
+        let d10hr_moisture = hco_x10hr(hour as i32, d1hr_results[hour], d1hr_6h, d1hr_15h);
+
+        *d10hr_value = d10hr_moisture;
+    }
+
+    for (hour, d100hr_value) in d100hr_results.iter_mut().enumerate() {
+        // For 100hr, apply the algorithm again to get the next size class
+        let d10hr_6h = d10hr_results[6];
+        let d10hr_15h = d10hr_results[15];
+        let d100hr_moisture = hco_x10hr(hour as i32, d10hr_results[hour], d10hr_6h, d10hr_15h);
+        *d100hr_value = d100hr_moisture;
+    }
+
+    HourlyMoistureResults {
+        d1hr: d1hr_results,
+        d10hr: d10hr_results,
+        d100hr: d100hr_results,
+    }
+}
 
 /// Returns the effect of a certain precipitation as a function
 /// of the month and how many days have passed since the day being
@@ -923,6 +1051,55 @@ mod tests {
                     }
                 }
             }
+        }
+    }
+
+    #[test]
+    fn test_hourly_moisture_calculation() {
+        let temperature = [15.0; 24];
+        let humidity = [60.0; 24];
+        let cloud_cover = [50.0; 24];
+        let slope = 0.577; // 30 degree slope as ratio (tan(30°))
+        let aspect = 3.14159; // 180 degrees as radians (π)
+        let precipitation_6_days = [0.0, 2.0, 0.0, 5.0, 0.0, 1.0];
+        let month = 6;
+
+        let results = calculate_hourly_fuel_moisture(
+            &temperature,
+            &humidity,
+            &cloud_cover,
+            slope,
+            aspect,
+            &precipitation_6_days,
+            month,
+            2,
+        );
+
+        // Basic sanity checks
+        assert_eq!(results.d1hr.len(), 24);
+        assert_eq!(results.d10hr.len(), 24);
+        assert_eq!(results.d100hr.len(), 24);
+
+        // All values should be non-negative
+        for i in 0..24 {
+            assert!(
+                results.d1hr[i] >= 0.0,
+                "d1hr[{}] = {} should be >= 0",
+                i,
+                results.d1hr[i]
+            );
+            assert!(
+                results.d10hr[i] >= 0.0,
+                "d10hr[{}] = {} should be >= 0",
+                i,
+                results.d10hr[i]
+            );
+            assert!(
+                results.d100hr[i] >= 0.0,
+                "d100hr[{}] = {} should be >= 0",
+                i,
+                results.d100hr[i]
+            );
         }
     }
